+++ celltk/core/operation.py
@@ -14,16 +14,16 @@ import skimage.morphology as morph
 import skimage.util as util
 import SimpleITK as sitk
 
-from cellst.core.arrays import ConditionArray
-from cellst.utils._types import (Image, Mask, Track, Arr, Same,
+from celltk.core.arrays import ConditionArray
+from celltk.utils._types import (Image, Mask, Track, Arr, Same,
                                  ImageContainer, INPT_NAMES,
                                  RandomNameProperty)
-from cellst.utils.operation_utils import (track_to_mask, parents_from_track,
+from celltk.utils.operation_utils import (track_to_mask, parents_from_track,
                                           match_labels_linear, mask_to_seeds)
-from cellst.utils.log_utils import get_console_logger
-from cellst.utils.utils import ImageHelper
-import cellst.utils.metric_utils as metric_utils
-import cellst.utils.filter_utils as filter_utils
+from celltk.utils.log_utils import get_console_logger
+from celltk.utils.utils import ImageHelper
+import celltk.utils.metric_utils as metric_utils
+import celltk.utils.filter_utils as filter_utils
 
 
 class Operation():
