+++ celltk/utils/kit_sch_ge_utils.py
@@ -4,7 +4,7 @@ from typing import Tuple
 import numpy as np
 import pandas as pd
 
-from cellst.utils._types import Image, Mask
+from celltk.utils._types import Image, Mask
 from kit_sch_ge.tracker.tracking import (TrackingConfig, MultiCellTracker,
                                          Tracklet)
 from kit_sch_ge.tracker.export import (ExportResults, catch_tra_issues,
