+++ celltk/utils/metric_utils.py
@@ -39,7 +39,7 @@ def cumulative_active(array: np.ndarray) -> np.ndarray:
 
 
 def predict_peaks(array: np.ndarray,
-                  weight_path: str = 'cellst/config/example_peak_weights.tf',
+                  weight_path: str = 'celltk/config/example_peak_weights.tf',
                   segment: bool = True,
                   ) -> None:
     """Placeholder function for now, but can change to include peaak prediction"""
