+++ celltk/utils/upeak/predict.py
@@ -1,0 +1,9 @@
+
+from celltk.utils.upeak.peak_utils import pad_traces
+
+
+def predict_peaks(arr: np.ndarray,
+                  ) -> np.ndarray:
+    """
+    """
+    arr = pad_traces(arr, )
