+++ requirements.txt
@@ -22,3 +22,4 @@ colorcet==3.0.0
 umap==0.1.1
 hdbscan==0.8.28
 seaborn==0.11.2
+pytest==7.0.1
